+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - Users can create and view Funds; users can create and view Activities
 - Users can download an XML representation of an Activity
 - Split activity form into multiple steps
+- Create Transactions associated with a Fund